BI-6174 Renaming processor package to consumer
