+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/Application.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import uk.gov.companieshouse.chipsrestinterfacesconsumer.processor.MessageConsumer;
+import uk.gov.companieshouse.chipsrestinterfacesconsumer.consumer.MessageConsumer;
 
 @SpringBootApplication
 public class Application implements CommandLineRunner {
