+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/consumer/MessageConsumer.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.chipsrestinterfacesconsumer.processor;
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.consumer;
 
 import uk.gov.companieshouse.chipsrestinterfacesconsumer.model.ChipsKafkaMessage;
 import java.util.Collection;
